Fix etter kommentarer på PR
