+++ src/main/web_src/src/components/fagsystem/sykdom/visning/Visning.tsx
@@ -6,40 +6,32 @@ import { DetaljertSykemelding } from './partials/DetaljertSykemelding'
 import { Sykemelding, SykemeldingSynt, SykemeldingDetaljert } from '../SykemeldingTypes'
 import { GyldigeBestillinger } from '~/components/transaksjonid/GyldigeBestillinger'
 
-type SykemeldingVisning = {
-	data: Sykemelding
-	ident: string
-	bestillingId: Array<number>
-}
-
-export const SykemeldingVisning = ({ data, ident, bestillingId }: SykemeldingVisning) => {
+export const SykemeldingVisning = ({ data }: Sykemelding) => {
 	// Viser foreløpig bestillingsdata
 	if (!data || data.length < 1) return null
-
-	const gyldigeBestillinger = GyldigeBestillinger(data, 'SYKEMELDING', ident)
 
 	return (
 		<div>
 			<SubOverskrift label="Sykemelding" iconKind="sykdom" />
-			{gyldigeBestillinger.map(
-				(bestilling: SykemeldingSynt | SykemeldingDetaljert, idx: number) => {
-					const syntSykemelding = _get(bestilling, 'data.sykemelding.syntSykemelding')
-					const detaljertSykemelding = _get(bestilling, 'data.sykemelding.detaljertSykemelding')
-
-					if (!bestilling.erGjenopprettet) {
-						return syntSykemelding ? (
-							<SyntSykemelding sykemelding={syntSykemelding} idx={idx} />
-						) : detaljertSykemelding ? (
-							<DetaljertSykemelding sykemelding={detaljertSykemelding} idx={idx} />
-						) : null
-					}
+			{data.map((bestilling: SykemeldingSynt | SykemeldingDetaljert, idx: number) => {
+				const syntSykemelding = _get(bestilling, 'data.sykemelding.syntSykemelding')
+				const detaljertSykemelding = _get(bestilling, 'data.sykemelding.detaljertSykemelding')
+				if (!bestilling.erGjenopprettet) {
+					return syntSykemelding ? (
+						<SyntSykemelding sykemelding={syntSykemelding} idx={idx} />
+					) : detaljertSykemelding ? (
+						<DetaljertSykemelding sykemelding={detaljertSykemelding} idx={idx} />
+					) : null
 				}
-			)}
+			})}
 		</div>
 	)
 }
 
-SykemeldingVisning.filterValues = (bestillinger: Array<Sykemelding>) => {
+SykemeldingVisning.filterValues = (bestillinger: Array<Sykemelding>, ident: number) => {
 	if (!bestillinger) return null
-	return bestillinger.filter((bestilling: any) => bestilling.data.sykemelding)
+	const sykemeldingBestillinger = bestillinger.filter(
+		(bestilling: any) => bestilling.data.sykemelding
+	)
+	return GyldigeBestillinger(sykemeldingBestillinger, 'SYKEMELDING', ident)
 }
