+++ src/main/web_src/src/pages/gruppe/PersonVisning/PersonVisning.js
@@ -51,9 +51,8 @@ export const PersonVisning = ({
 				ident={ident.ident}
 			/>
 			<SykemeldingVisning
-				data={SykemeldingVisning.filterValues(bestillingsListe)}
+				data={SykemeldingVisning.filterValues(bestillingsListe, ident.ident)}
 				ident={ident.ident}
-				bestillingId={ident.bestillingId}
 			/>
 			<BrregVisning data={data.brregstub} loading={loading.brregstub} />
 			<KrrVisning data={data.krrstub} loading={loading.krrstub} />
